API review: move avocado.exceptions.CmdError to avocado.utils.process.CmdError
This exception is only raised by avocado.utils.process itself (and by
avocado.utils.remote which mimics the the first). Also, since avocado.utils
should not use anything from avocado.core, it should be moved anyway.
